requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
binaryornot==0.4.4
boto3==1.34.127
botocore==1.34.127
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
crispy-bootstrap5==2024.2
cron-descriptor==1.4.5
cryptography==44.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.5.1
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-import-export==4.3.5
django-jet-reboot==1.3.7
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-ratelimit==4.1.0
django-recaptcha==3.0.0
django-ses==4.1.0
django-storages==1.14.3
django-thread==0.0.1
django-timezone-field==7.0
django-user-agents==0.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
easydjango-project==1.0.5
ecdsa==0.19.0
Faker==33.1.0
filelock==3.16.1
firebase-admin==6.5.0
flower==2.0.1
frozenlist==1.5.0
gevent==24.11.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
humanize==4.11.0
identify==2.6.3
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
MouseInfo==0.1.3
msgpack==1.1.0
multidict==6.1.0
nodeenv==1.9.1
oauthlib==3.2.2
openai==0.28.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pyperclip==1.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.3.0
python-pptx==0.6.23
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.0.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
rubicon-objc==0.4.9
s3transfer==0.10.4
scanreq==0.0.9
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.4.2
tablib==3.8.0
text-unidecode==1.3
tornado==6.4.2
tqdm==4.67.1
twilio==9.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2022.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
vine==5.1.0
virtualenv==20.28.0
wcwidth==0.2.13
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
